% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccination.R
\name{get_vaccination}
\alias{get_vaccination}
\title{Download vaccination data of Minsa Peru}
\usage{
get_vaccination(path = NULL, archive = NULL)
}
\arguments{
\item{path}{(Optional) You can specify where you want the data
to be downloaded.}

\item{archive}{(Optional) You can specify the location of the .7z
file you have previously downloaded so that the function can determine
if you have an outdated version and need to re-download a new version.}
}
\value{
An archive .7z and .csv file about vaccination data
}
\description{
The download source is MINSA's open data platform,
which updates covid data on an almost daily basis.
}
